Move the common ground bundle states to worker to make it accessible on both ends
