+++ codalab/bundles/uploaded_bundle.py
@@ -4,8 +4,8 @@ directly uploaded to the bundle store. Uploaded bundles are constructed with
 just a bundle store data hash and a metadata dict, and never need to be run.
 '''
 from codalab.bundles.named_bundle import NamedBundle
-from codalab.common import State
 from codalab.objects.metadata_spec import MetadataSpec
+from codalabworker.bundle_state import State
 
 class UploadedBundle(NamedBundle):
     METADATA_SPECS = list(NamedBundle.METADATA_SPECS)
