+++ codalab/lib/bundle_cli.py
@@ -42,7 +42,6 @@ from codalab.common import (
     NotFoundError,
     PermissionError,
     precondition,
-    State,
     UsageError,
 )
 from codalab.lib import (
@@ -92,6 +91,7 @@ from codalabworker.file_util import un_tar_directory
 from codalab.lib.spec_util import generate_uuid
 from codalabworker.docker_client import DockerClient
 from codalabworker.file_util import remove_path
+from codalabworker.bundle_state import State
 
 # Formatting Constants
 ADDRESS_SPEC_FORMAT = "(<alias>|<address>)"
