+++ worker/codalabworker/local_run/local_run_manager.py
@@ -244,7 +244,7 @@ class LocalRunManager(BaseRunManager):
                     'start_time': run_state.start_time,
                     'docker_image': run_state.docker_image,
                     'info': run_state.info,
-                    'state': LocalRunState.WORKER_STATE_TO_SERVER_STATE[run_state.stage]
+                    'state': LocalRunStage.WORKER_STATE_TO_SERVER_STATE[run_state.stage]
                 } for bundle_uuid, run_state in self.runs.items()
             }
             return result
